refactor(snap): Put install and configure hooks back
