+++ snap/local/hooks/configure.go
@@ -10,6 +10,7 @@
  * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
+ *
  * SPDX-License-Identifier: Apache-2.0'
  */
 
@@ -19,98 +20,27 @@ import (
 	"os"
 	"strings"
 
-	hooks "github.com/canonical/edgex-snap-hooks/v2"
-	"github.com/canonical/edgex-snap-hooks/v2/env"
 	"github.com/canonical/edgex-snap-hooks/v2/log"
 	"github.com/canonical/edgex-snap-hooks/v2/options"
 	"github.com/canonical/edgex-snap-hooks/v2/snapctl"
 )
 
-func main() {
-	subCommand := os.Args[1]
-	switch subCommand {
-	case "install":
-		install()
-	case "configure":
-		configure()
-	default:
-		panic("Unknown hook sub-command: " + subCommand)
-	}
-}
-
-// installProfiles copies the profile configuration.toml files from $SNAP to $SNAP_DATA.
-func installConfig() error {
-	path := "/config/device-virtual/res"
-
-	err := hooks.CopyDir(
-		env.Snap+path,
-		env.SnapData+path)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func installDevices() error {
-	path := "/config/device-virtual/res/devices"
-
-	err := hooks.CopyDir(
-		hooks.Snap+path,
-		hooks.SnapData+path)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func installDevProfiles() error {
-	path := "/config/device-virtual/res/profiles"
-
-	err := hooks.CopyDir(
-		hooks.Snap+path,
-		hooks.SnapData+path)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func install() {
-	log.SetComponentName("install")
-
-	err := installConfig()
-	if err != nil {
-		log.Fatalf("error installing config file: %s", err)
-	}
-
-	err = installDevices()
-	if err != nil {
-		log.Fatalf("error installing devices config: %s", err)
-	}
-
-	err = installDevProfiles()
-	if err != nil {
-		log.Fatalf("error installing device profiles config: %s", err)
-	}
-}
-
-//configure
+// configure is called by the main function
 func configure() {
 	log.SetComponentName("configure")
 
 	log.Info("Enabling config options")
 	err := snapctl.Set("app-options", "true").Run()
 	if err != nil {
-		log.Fatalf("could not enable config options: %v", err)
+		log.Errorf("could not enable config options: %v", err)
+		os.Exit(1)
 	}
 
 	log.Info("Processing options")
 	err = options.ProcessAppConfig("device-virtual")
 	if err != nil {
-		log.Fatalf("could not process options: %v", err)
+		log.Errorf("could not process options: %v", err)
+		os.Exit(1)
 	}
 
 	// If autostart is not explicitly set, default to "no"
@@ -118,7 +48,8 @@ func configure() {
 	// are provided by default.
 	autostart, err := snapctl.Get("autostart").Run()
 	if err != nil {
-		log.Fatalf("Reading config 'autostart' failed: %v", err)
+		log.Errorf("Reading config 'autostart' failed: %v", err)
+		os.Exit(1)
 	}
 	if autostart == "" {
 		log.Debug("autostart is NOT set, initializing to 'no'")
@@ -132,11 +63,13 @@ func configure() {
 	case "true", "yes":
 		err = snapctl.Start("device-virtual").Enable().Run()
 		if err != nil {
-			log.Fatalf("Can't start service: %s", err)
+			log.Errorf("Can't start service: %s", err)
+			os.Exit(1)
 		}
 	case "false", "no":
 		// no action necessary
 	default:
-		log.Fatalf("Invalid value for 'autostart': %s", autostart)
+		log.Errorf("Invalid value for 'autostart': %s", autostart)
+		os.Exit(1)
 	}
 }
